+++ lib/alacritty_themes.rb
@@ -2,12 +2,14 @@
 
 require "alacritty_themes/version"
 require_relative "alacritty_themes/parser"
+require_relative "alacritty_themes/file_helper"
 
 module AlacrittyThemes
   class Error < StandardError; end
 
-  # CLI entry point
-  class CLI
+  class CLI # :nodoc:
+    include FileHelper
+
     def start(argv)
       options = parse_options(argv)
       execute_command(options)
@@ -20,6 +22,11 @@ module AlacrittyThemes
     end
 
     def execute_command(options)
+      if options[:command] == :create
+        create_directory_path unless does_path_exist?
+        create_backup_file if does_file_exist?
+        create_file
+      end
       puts options[:message]
     end
   end