+++ spec/spec_helper.rb
@@ -2,6 +2,7 @@
 
 require "bundler/setup"
 require "alacritty_themes"
+require "fakefs/spec_helpers"
 
 RSpec.configure do |config|
   # Enable flags like --only-failures and --next-failure
@@ -13,4 +14,5 @@ RSpec.configure do |config|
   config.expect_with :rspec do |c|
     c.syntax = :expect
   end
+  config.include FakeFS::SpecHelpers
 end